Connection integrity improvements, alert improvements, GUI and screen code improvements.
Connecting is now attempted every 5 seconds untill a connection is made.
Adding a contact, logging out, logging in, creating an account; are all
handled when connection has been lost.
Alerts are centered.
Creating an account that already exists will alert the user of that.
All Screens have access to loadLater and alert Later.