Display connection. General GUI improvements.
Connection is now displayed. a forced check can be performed by clicking
on it.
The screen size is locked to a good size, and all elements fit nicely
inside.
Only relevant text boxes are in focus, nothing else can be focused on.
Pressing enter sends an action, escape exits it.
A contact request cannot be shown for a contact that is either already
your contact, or already has a request in your list.